Move frozen modules into their own directory.
