+++ Makefile.pre.in
@@ -748,25 +748,25 @@ regen-frozen: Tools/scripts/freeze_modules.py $(FROZEN_FILES)
 
 # BEGIN: freezing modules
 
-Python/frozen_importlib__bootstrap.h: $(srcdir)/Programs/_freeze_module $(srcdir)/Lib/importlib/_bootstrap.py
+Python/frozen_modules/importlib__bootstrap.h: $(srcdir)/Programs/_freeze_module $(srcdir)/Lib/importlib/_bootstrap.py
 	$(srcdir)/Programs/_freeze_module importlib._bootstrap \
 		$(srcdir)/Lib/importlib/_bootstrap.py \
-		$(srcdir)/Python/frozen_importlib__bootstrap.h
+		$(srcdir)/Python/frozen_modules/importlib__bootstrap.h
 
-Python/frozen_importlib__bootstrap_external.h: $(srcdir)/Programs/_freeze_module $(srcdir)/Lib/importlib/_bootstrap_external.py
+Python/frozen_modules/importlib__bootstrap_external.h: $(srcdir)/Programs/_freeze_module $(srcdir)/Lib/importlib/_bootstrap_external.py
 	$(srcdir)/Programs/_freeze_module importlib._bootstrap_external \
 		$(srcdir)/Lib/importlib/_bootstrap_external.py \
-		$(srcdir)/Python/frozen_importlib__bootstrap_external.h
+		$(srcdir)/Python/frozen_modules/importlib__bootstrap_external.h
 
-Python/frozen_zipimport.h: $(srcdir)/Programs/_freeze_module $(srcdir)/Lib/zipimport.py
+Python/frozen_modules/zipimport.h: $(srcdir)/Programs/_freeze_module $(srcdir)/Lib/zipimport.py
 	$(srcdir)/Programs/_freeze_module zipimport \
 		$(srcdir)/Lib/zipimport.py \
-		$(srcdir)/Python/frozen_zipimport.h
+		$(srcdir)/Python/frozen_modules/zipimport.h
 
-Python/frozen_hello.h: $(srcdir)/Programs/_freeze_module $(srcdir)/Tools/freeze/flag.py
+Python/frozen_modules/hello.h: $(srcdir)/Programs/_freeze_module $(srcdir)/Tools/freeze/flag.py
 	$(srcdir)/Programs/_freeze_module hello \
 		$(srcdir)/Tools/freeze/flag.py \
-		$(srcdir)/Python/frozen_hello.h
+		$(srcdir)/Python/frozen_modules/hello.h
 
 # END: freezing modules
 
@@ -1001,10 +1001,10 @@ Python/ceval.o: $(srcdir)/Python/opcode_targets.h $(srcdir)/Python/ceval_gil.h \
 
 # FROZEN_FILES is auto-generated by Tools/scripts/freeze_modules.py.
 FROZEN_FILES = \
-		$(srcdir)/Python/frozen_importlib__bootstrap.h \
-		$(srcdir)/Python/frozen_importlib__bootstrap_external.h \
-		$(srcdir)/Python/frozen_zipimport.h \
-		$(srcdir)/Python/frozen_hello.h
+		$(srcdir)/Python/frozen_modules/importlib__bootstrap.h \
+		$(srcdir)/Python/frozen_modules/importlib__bootstrap_external.h \
+		$(srcdir)/Python/frozen_modules/zipimport.h \
+		$(srcdir)/Python/frozen_modules/hello.h
 # End FROZEN_FILES
 
 Python/frozen.o: $(FROZEN_FILES)
