+++ Python/frozen_modules/importlib__bootstrap_external.h
